Fleshed out entities a little bit
The Squarelet entity was amended, to contain the actual binary data,
as also the replies for the Squarelet. SquareletService was replaced
with a more generic posting service, which will contain the ACL
checking for posting new content in the future.

The UserAccount entity now holds a String as it's primary key, which
comes in handy, when deploying on AppEngine. To accomodate for some
DataNucleus hickups, an GAE specific aspect was written, that enhances
the UserAccount entity with DataNucleus specific extension annotations
for GAE.