UserAccount should implement UserDetails
The UserAccount class will later on be handed over to spring-security
for role assessment. Therefore it now implements the UserDetails
interface. Also the identifier field has been demoted to a normal
member of the class. Using a DataNucleus extension, it is used to derive
the actual id.

In the wake the new Roles enum is introduced, which distinguishes between
new users, registered users and users with admin rights.